requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary
python-dotenv>=1.0.0
python-multipart>=0.0.9
requests
pydantic>=2.0.0

# Layer1 - Complaint Dataset Generator Agent
google-generativeai>=0.3.0
